requirement.txt
minium==1.4.1
tidevice==0.10.8
pyYAML==6.0